pyproject.toml
[project]
name = "log-playground"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "logfire[fastapi]>=4.13.2",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.14.0",
]

[tool.ruff]
output-format = "concise"

[tool.ruff.lint]
select = [
    "E", "W",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
    "PL",  # Pylint
    "RUF",
]

[tool.mypy]
strict = true
plugins = ['pydantic.mypy']
